Pass updated layer_metadata_map from Model Transformer
ModelTransformer gets layer_metadata_map which contains quant
information for layers. The transforms have the ability to update
this metadata but this wasn't being passed back and used.

Now updated quant metadata is used in quantization and completes
the loop. Without any transform modifying it, there won't be any
difference to the execution path.

PiperOrigin-RevId: 286038157